+++ packages/yak_flutter/CHANGELOG.md
@@ -1,8 +1,10 @@
-### 1.0.2+0
+### 1.0.3+0
+- update dependencies
+
+### 1.0.2
 - breaking changes in `yak_runner`
 
-### 1.0.1+2
-- sync version with `yak_runner`
+### 1.0.1
 - update dependencies
 
 ### 1.0.0