+++ packages/yak_tween/CHANGELOG.md
@@ -1,4 +1,4 @@
-### 0.1.0+13
+### 0.1.0+14
 - fix analysis
 - updated `example`
 - created from the ashes of `yak_math`